fixed: stream starts up in the middle of the call
Bug description: On the iOS app now, if the stream is playing and the phone rings, the stream starts up again in the middle of the call. This will not work for obvious reasons! Can you look at the app again and see if there is a way to have the stream stop if there is a call?